mime-types(mime-add): Add.
It's often the case that the mime DB provided by trivial-mimes or OS
is terribly incomplete. Until this PR, there was not reliable way to
add MIMEs to trivial-mimes DB but modifying the *mime-db*
and *reverse-mime-db* (unexported, thus discouraged) variables.

This PR adds mime-add, a function to add a MIME and its corresponding
file extensions to both the DBs trivial-mimes has. Maybe no need to
export *mime-db* then?